Move getNoRowMessage to common results utils directory
