+++ src/features/results/utils/index.ts
@@ -6,3 +6,4 @@
  */
 
 export * from './renderTable-ExportCsv';
+export * from './no-rows-message';
